backend/app/db/mongo.py
# Future place for database models
import os
from dotenv import load_dotenv
from motor.motor_asyncio import AsyncIOMotorClient

load_dotenv()
MONGO_URI = os.getenv("MONGO_URI")

client = AsyncIOMotorClient(MONGO_URI)
db = client["gymbuddy"]
chatlogs = db["chatlogs"]